refactor: InputView와 ResultView를 나누는 작업
