+++ src/main/kotlin/minesweeper/view/InputView.kt
@@ -2,5 +2,19 @@ package minesweeper.view
 
 object InputView {
 
-    fun receiptInt() = readLine()!!.toInt()
+    fun printToReceiptVertical() {
+        println("높이를 입력하세요.")
+    }
+
+    fun printToReceiptHeight() {
+        println("세로를 입력하세요.")
+    }
+
+    fun printToReceiptCountOfMine() {
+        println("지뢰는 몇 개인가요?")
+    }
+
+    fun printToStartGame() {
+        println("지뢰찾기 게임 시작")
+    }
 }
