+++ src/main/kotlin/minesweeper/view/Reception.kt
@@ -1,0 +1,6 @@
+package minesweeper.view
+
+object Reception {
+
+    fun receiptInt() = readLine()!!.toInt()
+}
