+++ src/main/kotlin/minesweeper/view/ResultView.kt
@@ -2,7 +2,7 @@ package minesweeper.view
 
 object ResultView {
 
-    const val START_ROW = 0
+    private const val START_ROW = 0
 
     fun printAllMineGround(markersWithMine: MutableList<List<String>>) {
         (markersWithMine.indices).map { x ->
@@ -15,21 +15,5 @@ object ResultView {
         (markersWithMine[START_ROW].indices).map { row ->
             print(markersWithMine[column][row])
         }
-    }
-
-    fun printToReceiptVertical() {
-        println("높이를 입력하세요.")
-    }
-
-    fun printToReceiptHeight() {
-        println("세로를 입력하세요.")
-    }
-
-    fun printToReceiptCountOfMine() {
-        println("지뢰는 몇 개인가요?")
-    }
-
-    fun printToStartGame() {
-        println("지뢰찾기 게임 시작")
     }
 }
